Detalles finales, validaciones y otros

+++ Consola/funciones.py
@@ -5,27 +5,62 @@ from colorama import *
 import time
 import os
 
+def valDate(date):
+    try:
+        datetime.strptime(date, '%d-%m-%y')
+        return True
+    except ValueError:
+        return False
+    
+def valInt(num):
+    try:
+        int(num)
+        IaN = True
+    except ValueError:
+        IaN = False
+    if IaN == True:
+        return num
+    else:
+        print("Valor no valido, intente de nuevo... \n Valor esperado: Número entero")
+        return valInt(int(input()))
+
 taskList = []
 estado=""
 
+def detectFile():
+    if os.path.isfile("tareas.txt"):
+        print("Archivo detectado")
+    else:
+        print("Archivo no detectado, creando archivo...")
+        file = open("tareas.txt", "w")
+        file.close()
+
 def writeFile(taskList):
-    file = open("tareas.txt", "w")
+    file = open("tareas.txt", "w", encoding='utf-8')
+    for i in taskList:
+        file.write(i.nombre + "," + i.detalles + "," + str(i.fechaC) + "," + i.fechaL + "," + i.estado + "\n")
     file.close()
 
 def readFile():
-    file = open("tareas.txt", "r")
-    data = file.readlines()
-    for d in data:
-        nombre, detalles, fechaC, fechaL, estado = d.rstrip.split(",")
-        task = tarea(nombre, detalles, fechaC, fechaL, estado)
-        taskList.append(task)
-    file.close()
+    if os.path.isfile("tareas.txt"):
+        with open("tareas.txt", "r", encoding='utf-8') as file:
+            lines = file.readlines()
+            for line in lines:
+                parts = line.strip().split(",")
+                if len(parts) == 5:  # Check if line has enough parts
+                    task = tarea(parts[0], parts[1], parts[2], parts[3], parts[4])
+                    taskList.append(task)
 
 def add_tarea():
     nombre = input("Nombre de la tarea: ")
     detalles = input("Detalles de la tarea: ")
-    fechaC = datetime.now()
-    fechaL = input("Fecha límite de la tarea (dd-mm-yyyy): ")
+    fechaC = datetime.now().strftime("%d-%m-%Y")
+    fechaL = input("Fecha límite de la tarea (Formato: dd-mm-yyyy) \n Para no ingresar fecha límite presione enter:")
+    if fechaL == "":
+        fechaL = "Sin fecha límite"
+    elif fechaL != valDate(fechaL):
+        print("Fecha no válida")
+        fechaL = input("Fecha límite de la tarea (dd-mm-yyyy): ")
     estado = "pendiente"
     task = tarea(nombre, detalles, fechaC, fechaL, estado)
     return task
@@ -34,15 +69,61 @@ def show_tareas(taskList):
     data = []
     for i in range(len(taskList)):
         data.append([taskList[i].nombre, taskList[i].detalles, taskList[i].fechaC, taskList[i].fechaL, taskList[i].estado])
-        print(tabulate(data, headers=["Nombre", "Detalles", "Fecha Creación", "Fecha Límite", "Estado"],showindex="always",
+    print(tabulate(data, headers=["Nombre", "Detalles", "Fecha Creación", "Fecha Límite", "Estado"],showindex="always",
                        tablefmt="fancy_grid", stralign="center", numalign="center"))
 
-def filtrar_por_estado(taskList, estado):
+def filtrar_por_estado(taskList):
+    data = []
+    print("1. Pendiente")
+    print("2. En proceso")
+    print("3. Terminado")
+    print("Presione cualquier tecla para cancelar")
+    print("Opción:", end=" ")
+    opcion = valInt(int(input()))
+    if opcion == 1:
+        estado = "pendiente"
+    elif opcion == 2:
+        estado = "en proceso"
+    elif opcion == 3:
+        estado = "terminado"
+    else:
+        print("Cancelando...")
+        return
     for i in taskList:
         if i.estado.lower() == estado.lower():
-            print(i.nombre, i.detalles, i.fechaC, i.fechaC, i.fechaL, i.estado, "\n")
+            data.append([i.nombre, i.detalles, i.fechaC, i.fechaL, i.estado])
+    print(tabulate(data, headers=["Nombre", "Detalles", "Fecha Creación", "Fecha Límite", "Estado"],showindex="always",
+                   tablefmt="fancy_grid", stralign="center", numalign="center"))
 
-
+def changeState(estado):
+    if(estado == "pendiente"):
+        print("Estado Actual: ", estado)
+        print("1. En proceso")
+        print("2. Terminado")
+        print("Presione cualquier tecla para cancelar")
+        print("Opción:", end=" ")
+        opcion = valInt(int(input()))
+        if(opcion == 1):
+            return "en proceso"
+        elif(opcion == 2):
+            return "terminado"
+        else:
+            return estado
+    elif(estado == "en proceso"):
+        print("Estado Actual: ", estado)
+        print("1. Pendiente")
+        print("2. Terminado")
+        print("Presione cualquier tecla para cancelar")
+        opcion = int(input("Opción: "))
+        if(opcion == 1):
+            return "pendiente"
+        elif(opcion == 2):
+            return "terminado"
+        else:
+            return estado
+    elif(estado == "terminado"):
+        print("Las tareas terminadas ya no pueden editarse")
+        return estado        
 
 def mod_tarea(taskList):
     nombre = input("Nombre de la tarea a modificar: ")
@@ -61,7 +142,7 @@ def mod_tarea(taskList):
             elif opcion == 3:
                 i.fechaL = input("Nueva fecha límite: ")
             elif opcion == 4:
-                i.estado = input("Nuevo estado (En Progreso, Finalizada, Pendiente): ")
+                i.estado = changeState(i.estado)
             else:
                 print("Cancelar modificación")
                 break
@@ -77,7 +158,7 @@ def del_tarea(taskList):
         else:
             print("Tarea no encontrada")
 
-def main():
+def menu():
     readFile()
     while True:
         print("1. Agregar tarea")
@@ -86,7 +167,8 @@ def main():
         print("4. Eliminar tarea")
         print("5. Filtrado")
         print("6. Salir")
-        opcion = int(input("Opción: "))
+        print("Opción:", end=" ")
+        opcion = valInt(int(input()))
         if opcion == 1:
             os.system("cls")
             taskList.append(add_tarea())
@@ -109,17 +191,29 @@ def main():
             os.system("cls")
         elif opcion == 5:
             os.system("cls")
-            estado=input("Ingrese el estado que desea ver ")
-            filtrar_por_estado(taskList, estado)
+            filtrar_por_estado(taskList)
             input("Presione enter para continuar...")
             os.system("cls")
         elif opcion == 6:
-            
-            writeFile(taskList)
-            break
+            exitApp(taskList)
         else:
             print("Opción no válida, intente de nuevo...")
             input("Presione enter para continuar...")
+
+def exitApp(taskList):
+    if(int(input("¿Desea salir de la aplicacion? \n 1. Salir \n 2. Cancelar \n ")) == 1):
+        print("Saliendo...")
+        writeFile(taskList)
+        time.sleep(2)
+        print("Archivo guardado, gracias por usar el programa...")
+        time.sleep(2)
+        os.system("cls")
+        quit()
+    else:
+        print("Cancelando...")
+        time.sleep(2)
+        os.system("cls")
+        input("Presione enter para continuar...")
 
 def pantallaInicio():
     init()
@@ -145,4 +239,5 @@ def pantallaInicio():
 
 
 pantallaInicio()
-main()
+detectFile()
+menu()
